Added Envelope Implementation
To overcome https://github.com/imiskolee/FormUpload/issues/2 which is why I've abandonned the legacy code-base, I worked out that you'd need something similar to html, with nodes, of-which some could contain other nodes in specific types. This is the Envelope type (I think I saw the word envelope in the RFC), and it's designed to hold other nodes. Unlike HTML, this will probably be the only type of node to hold other nodes. For now I've avoided array access, other such niceness just to get it working